+++ src/api/v1/apps/main/__init__.py
@@ -1,0 +1,6 @@
+"""
+This app adresses all related with the main page such as
+- Dashboards
+- Menus
+- Statitistics
+"""