+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This subversion adds the financial dashborad entire creation
 
 ### Added
 - Template for financial dahboard in `src/templates/main/financial_dashboard.html`
+- New app called main for address dashboards and statistics in `src/api/v1/apps/main/`
 
 ### Changed
 - N/A
@@ -33,7 +34,7 @@ This subversion adds the financial dashborad entire creation
 This patch fix the total flow in the UI
 
 ### Added
-- N/A
+- Router for financial dashboard in `src/api/v1/apps/main/router.py`
 
 ### Changed
 - Added logic for no pending waiter orders in `src/templates/orders/waiter/cards.html`